@@ -1,17 +1,23 @@
 <script lang="ts">
   import { closeModal } from 'svelte-modals';
 
-  // provided by Modals
-  export let isOpen: boolean;
+  
+  interface Props {
+    // provided by Modals
+    isOpen: boolean;
+    children?: import('svelte').Snippet;
+  }
+
+  let { isOpen, children }: Props = $props();
 </script>
 
 {#if isOpen}
   <div role="dialog" class="modal">
     <div class="actions">
-      <button id="x" on:click={closeModal}>✕</button>
+      <button id="x" onclick={closeModal}>✕</button>
     </div>
     <div class="contents">
-      <slot />
+      {@render children?.()}
     </div>
   </div>
 {/if}
